% Generated by roxygen2 (4.0.1): do not edit by hand
\name{defactorize}
\alias{defactorize}
\title{Change factors columns to character}
\usage{
defactorize(df)
}
\arguments{
\item{df}{the \code{data.frame} to manipulate.}
}
\value{
a \code{data.frame} with characters vectors instead of factors
}
\description{
This function changes factors columns in a \code{data.frame} to character.
}

