% Generated by roxygen2 (4.0.1): do not edit by hand
\name{save.xlsx}
\alias{save.xlsx}
\title{save a bunch of objects in a xlsx file}
\usage{
save.xlsx(file, ...)
}
\arguments{
\item{file}{path of the xlsx file}

\item{...}{objects to be saved}
}
\value{
nothing
}
\description{
This function saves the objects in '...' argument to the 'file' xlsx file
 one object per sheet.
}
\details{
'file' path is passed verbatim to java so no bash style interpolation (eg '~')
 is performed
}
\seealso{
\code{write.xlsx}
}

